Various Changes
Added a table of current active cops to client side police script. Made it so `DutyStatus()` can be called with a player's local client ID as an argument to check their law enforcement status. This allows other scripts (like the wanted script) to check if a player is a cop. Also made some fixes to the police duty scripts server side.