Login Script Works
Made a lot of changes to the login script. It should be fully functional as of now. You can now play the game either based on a FiveM license ID, or a Steam ID. The game will prefer a Steam ID over a FiveM ID when determining how to load a character. Also updated the SQL file. It can be imported OVER the existing database. However, this version of the database script will delete any existing tables due to datatype conflicts.