Login Updates
Worked on some login updates for creating player accounts and such. Also removed the dependency on Steam. For people who still want to use Steam, the script is set up to use Steam as the primary identifier, but it's no longer required. Order of precedence for login is Steam, Social Club, FiveM, Discord (id)