Vehicles: Lock/Window Smash (See Notes)
Added an initial vehicle script. Added vehicle locks and window smashing, tested and working. Unoccupied vehicles have a chance to be unlocked when a player tried to get inside of it. Occupied vehicles are always locked. Added a window breaker item that allows players to smash the window of a locked vehicle (occupied or not).